Added Uniforms for DirectX10
Added uniforms for DirectX10. Need to do some work on automatically transposing matrices depending on the API. For some reason a GLM orthographic view matrix only works with OpenGL, but an XMMatrix orthographic view matrix works with both OpenGL and DirectX10.

Added some stuff to ResourceLib to allow for selecting a resource based the rendering mode. It's far more generic than that, but that's the original purpose I built it for.

+++ tau/TauEngine/src/camera/Camera2D.cpp
@@ -23,7 +23,7 @@ void Camera2D::setProjection(const Window& window, float bottom, float left) noe
 
 void Camera2D::setProjection(float top, float right, float bottom, float left) noexcept
 {
-    _projectionMatrix = glm::ortho(left, static_cast<float>(right), bottom, static_cast<float>(top), -1.0f, 1.0f);
+    _projectionMatrix = glm::ortho(left, right, bottom, top, -1.0f, 1.0f);
     _compoundedMatrix = _projectionMatrix * _viewMatrix;
 }
 
@@ -35,6 +35,7 @@ void Camera2D::recomputeMatrices() noexcept
     const glm::mat4 transform = glmExt::translate(identity, _position) * 
                                 glmExt::rotateDegrees(identity, _rotation, axis);
     _viewMatrix = glm::inverse(transform);
+    // _viewMatrix = transform;
     _compoundedMatrix = _projectionMatrix * _viewMatrix;
 }
